refactor(database/migrations): change gaji_awal and gaji_akhir data type to double
The data type for gaji_awal and gaji_akhir in the lowongan_kerja table was changed from decimal to double to simplify the schema and reduce unnecessary precision constraints. Additionally, indentation in the LowonganKerjaResource form was adjusted for better readability.

+++ database/migrations/2025_05_03_055749_create_lowongan_kerja_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->string('nama');
             $table->string('slug');
             $table->text('persyaratan'); 
-            $table->decimal('gaji_awal', 15, 2)->nullable();
-            $table->decimal('gaji_akhir', 15, 2)->nullable();
+            $table->double('gaji_awal')->nullable();
+            $table->double('gaji_akhir')->nullable();
             $table->boolean('tampilkan_gaji')->default(false);
             $table->enum('status', ['Draft', 'Publish', 'Arsip']);
             $table->enum('status_loker', ['Urgent', 'Normal', 'Full']);
